refactor(uniquely/salavat-pwa/alwatr-pwa): file structure
